refactor(ui): consolidate UI components into main components directory
- Moved all in-use components from ui/src/components/ui to ui/src/components
- Updated component filenames to follow PascalCase naming convention
- Merged button implementation into existing Button.tsx
- Created BaseInput.tsx to avoid conflict with existing Input.tsx
- Deleted unused skeleton component from ui directory
- Updated all import references throughout the codebase
- Removed the now-empty ui/src/components/ui directory

This refactoring improves code organization by eliminating the redundant
ui subdirectory and consolidating all components in one location with
consistent naming conventions.
