+++ ui/src/components/BaseInput.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { cn } from '~/components/utils';
 
-function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
+function BaseInput({ className, type, ...props }: React.ComponentProps<'input'>) {
   return (
     <input
       type={type}
@@ -18,4 +18,4 @@ function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
   );
 }
 
-export { Input };
+export { BaseInput };
