+++ ui/src/components/Skeleton.tsx
@@ -1,3 +1,15 @@
-import { Skeleton } from './ui/skeleton';
+import { cn } from '~/components/utils';
+
+function Skeleton({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) {
+  return (
+    <div
+      className={cn('animate-pulse rounded-md bg-muted', className)}
+      {...props}
+    />
+  );
+}
 
 export { Skeleton };
