+++ ui/src/components/environments/branches/BranchActionsDropdown.tsx
@@ -18,7 +18,7 @@ import {
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger
-} from '~/components/ui/dropdown-menu';
+} from '~/components/DropdownMenu';
 
 import { IEnvironment, ProposalState, SCM } from '~/types/Environment';
 import { Product } from '~/types/Meta';
