+++ ui/src/components/environments/branches/CreateMergeProposalModal.tsx
@@ -17,7 +17,7 @@ import {
   DialogTitle
 } from '~/components/Dialog';
 import Loading from '~/components/Loading';
-import { Button } from '~/components/ui/button';
+import { Button } from '~/components/Button';
 
 import { IEnvironment } from '~/types/Environment';
